requirements.txt
streamlit
google-genai
python-dotenv
langchain
PyPDF2
chromadb